app/requirements.txt
diffusers==0.24.0
ffmpeg_python==0.2.0
moviepy==2.0.0.dev0
numpy==1.26.2
openai_whisper==20231117
opencv_contrib_python==4.8.1.78
Pillow==10.1.0
Pillow==10.1.0
python-dotenv==1.0.0
Requests==2.31.0
runpod==1.4.2
torch==2.1.2
